extracting additional mapping processors to split responsibility between classes
