+++ modules/core/src/com/haulmont/cuba/core/sys/persistence/mapping/processors/SoftDeleteJoinExpressionProvider.java
@@ -14,11 +14,10 @@
  * limitations under the License.
  */
 
-package com.haulmont.cuba.core.sys.persistence.mapping.processors.softdelete;
+package com.haulmont.cuba.core.sys.persistence.mapping.processors;
 
 import com.google.common.base.Strings;
 import com.haulmont.cuba.core.entity.SoftDelete;
-import com.haulmont.cuba.core.sys.persistence.mapping.processors.AbstractJoinExpressionProvider;
 import org.apache.commons.lang3.reflect.FieldUtils;
 import org.eclipse.persistence.descriptors.ClassDescriptor;
 import org.eclipse.persistence.expressions.Expression;
